Feat : 평가 받은 회원의 경험치 추가

+++ src/main/java/com/spacestar/back/rate/domain/Experience.java
@@ -25,9 +25,9 @@ public class Experience extends GlobalTime {
 	private String uuid;
 
 	@Column(length = 10)
-	private Long changeExp;
+	private short changeExp;
 
-	public Experience(Long id, String uuid, Long changeExp) {
+	public Experience(Long id, String uuid, short changeExp) {
 		this.id = id;
 		this.uuid = uuid;
 		this.changeExp = changeExp;
